requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql==0.0.2
mysqlclient==2.0.3
numpy==1.20.2
pandas==1.2.4
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.4.11
tqdm==4.60.0
urllib3==1.26.4
webdriver-manager==3.4.1
Werkzeug==1.0.1
